+++ src/rh/tools/signal_detection.py
